+++ sources/packages/backend/apps/api/src/route-controllers/application-offering-change-request/models/application-offering-change-request.dto.ts
@@ -118,3 +118,27 @@ export class CreateApplicationOfferingChangeRequestAPIInDTO {
   @MaxLength(REASON_MAX_LENGTH)
   reason: string;
 }
+
+/**
+ * All in progress application offering change details.
+ */
+export class AllInProgressApplicationOfferingChangesAPIOutDTO {
+  id: number;
+  applicationNumber: string;
+  applicationId: number;
+  studyStartDate: string;
+  studyEndDate: string;
+  fullName: string;
+  status: ApplicationOfferingChangeRequestStatus;
+  createdAt: string;
+  studentId: number;
+}
+
+/**
+ * Extended pagination option for the in progress application offering change.
+ */
+export class AllInProgressOfferingChangePaginationOptionsAPIInDTO extends PaginationOptionsAPIInDTO {
+  @IsOptional()
+  @IsIn(["dateSubmitted", "status"])
+  sortField?: string;
+}
